// src/game-server/Enums/MapID.ts
export enum MapID {
	AshesPits1 = 37,
	AshesPits2 = 38,
	AshesPits3 = 39,
	AshesPits4 = 40,
	Badlands = 15,
	BadlandsVastMine = 88,
	BattleGrounds = 110,
	Blython = 2,
	BlythonArmory = 76,
	BlythonBar = 46,
	BlythonCityHall = 45,
	BlythonChurch = 106,
	BlythonGrocery = 48,
	BlythonHerbalShop1 = 47,
	BlythonHerbalShop2 = 78,
	BlythonMWBank = 51,
	BlythonSkillCenter = 107,
	BlythonStuffShop = 50,
	BlythonTradingZone = 52,
	BlythonWeaponry1 = 49,
	BlythonWeaponry2 = 77,
	BoneDesert = 16,
	BoneDesertNarrowMine = 89,
	BoneDesertVastMine = 90,
	Burial1 = 22,
	Burial2 = 24,
	Burial3 = 25,
	Burial4 = 26,
	Burial5 = 27,
	BurialRuins = 23,
	CursedAbyss = 18,
	CursedAbyssNarrowMine = 43,
	DemonLair1 = 30,
	DemonLair2 = 31,
	DemonLair3 = 32,
	DemonLair4 = 33,
	DemonSquare = 13,
	DemonSquareNarrowMine = 41,
	DemonSquarePrison = 67,
	DemonSquareCityHallFloor1 = 65,
	DemonSquareCityHallFloor2 = 66,
	DesertCity = 12,
	DesertCityCityHallFloor1 = 54,
	DesertCityCityHallFloor2 = 55,
	DesertCityNarrowMine = 86,
	DesertCityPrison = 72,
	DevilsGate = 21,
	DevilsGateNarrowMine = 42,
	EvilLair1 = 68,
	EvilLair2 = 69,
	EvilLair3 = 70,
	EvilLair4 = 71,
	FlameRuins = 19,
	FlameRuinsVastMine = 44,
	Flamingor = 20,
	GuildHQ = 100,
	OutcastCity = 14,
	OutcastCityMine = 87,
	ReviveArena = 3,
	Sewer1 = 34,
	Sewer2 = 35,
	Sewer3 = 36,
	ShowCase = 109,
	SkyPassage = 7,
	SunsetPlain = 17,
	TreeofLife = 8,
	TreeofLife2 = 9,
	TreeofLife3 = 10,
	TreeofLife4 = 11,
	TreeofLife5 = 73,
	TreeofLife6 = 74,
	TreeofLife7 = 75,
	Wasteland = 27,
	Wasteland1 = 28,
	Wasteland2 = 29,
	Wasteland3 = 105,
	Woodlingor = 1,
	WoodlingorArmory = 64,
	WoodlingorBar = 58,
	WoodlingorEldersZone = 4,
	WoodlingorGrocery = 59,
	WoodlingorHerbalShop = 60,
	WoodlingorMWBank = 61,
	WoodlingorSkillCenter = 63,
	WoodlingorStuffShop = 5,
	WoodlingorTradingZone = 62,
	WoodlingorWeaponry = 6,
}